refactor: EventDescription
