+++ shared/components/common/heartButton.tsx
@@ -4,16 +4,16 @@ import Button from "./button";
 
 
 const HeartButton = (props: { person: any; }) => {
-    const [person, setPerson] = useState(props.person)
+    const [active , setActive ] = useState(props.person)
     return (
         <Button
-            className={`h-10 w-10 rounded-full hover:scale-110 ${person ? "bg-red-600 dark:bg-red-600" : "" }`}
+            className={`h-10 w-10 rounded-full hover:scale-110 ${active ? "bg-red-600 dark:bg-red-600" : "" }`}
             onClick={() => {
-                setPerson(!person)
+                setActive (!active)
             }}
-            title={person ? 'Удалить из избранного' : 'Добавить в избранное'}
+            title={active ? 'Удалить из избранного' : 'Добавить в избранное'}
         >
-            <HeartIcon className={`stroke-0 ${person ? "fill-cd_main" : "dark:fill-cd_main"}`}/>
+            <HeartIcon className={`stroke-0 ${active ? "fill-cd_main" : "dark:fill-cd_main"}`}/>
 
         </Button>
     )
