+++ shared/components/timetable/viewMapCard.tsx
